Fixed SD2-826: Excluding folder from project browser throws exception
git-svn-id: svn://svn.sharpdevelop.net/sharpdevelop/branches/2.0@1485 1ccf3a8d-04fe-1044-b7c0-cef0b8235c61

+++ src/Main/Base/Project/Src/Gui/Pads/ProjectBrowser/Commands/DefaultFileNodeCommands.cs
@@ -61,7 +61,9 @@ namespace ICSharpCode.SharpDevelop.Project.Commands
 			dependentNodes.ForEach(ExcludeFileNode);
 			
 			bool isLink = fileNode.IsLink;
-			ProjectService.RemoveProjectItem(fileNode.Project, fileNode.ProjectItem);
+			if (fileNode.ProjectItem != null) { // don't try to exclude same node twice
+				ProjectService.RemoveProjectItem(fileNode.Project, fileNode.ProjectItem);
+			}
 			if (isLink) {
 				fileNode.Remove();
 			} else {
